shop

+++ internet_shop/src/main/java/ru/job4j/shop/controller/ProductsController.java
@@ -33,10 +33,45 @@ public class ProductsController extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        String action = req.getParameter("action");
-        String id = req.getParameter("id");
-        if ("delete".equals(action)) {
-            service.deleteProduct(id);
+        if (req.getSession().getAttribute("admin") != null) {
+            String action = req.getParameter("action");
+            String id = req.getParameter("id");
+            if ("delete".equals(action)) {
+                service.deleteProduct(id);
+            } else {
+                String form = req.getParameter("form");
+                String[]f = form.split("&");
+                String name = f[0].split("=")[1];
+                String price = f[1].split("=")[1];
+                String amount = f[2].split("=")[1];
+                String minidescription = f[3].split("=")[1].replaceAll("\\+", " ");
+                String description = f[4].split("=")[1].replaceAll("\\+", " ");
+                String result = "";
+                if (name.isEmpty() || price.isEmpty() || amount.isEmpty() || minidescription.isEmpty() || description.isEmpty()) {
+                    result = "Заполните все поля";
+                } else {
+                    Product product;
+                    try {
+                        int priceInt = Integer.parseInt(price);
+                        int amountInt = Integer.parseInt(amount);
+                        product = new Product(name, description, minidescription, priceInt, amountInt);
+                        if ("add".equals(action)) {
+                            service.addNewProduct(product);
+                            result = "ok";
+                        }
+                        if ("update".equals(action)) {
+                            service.updateProduct(product, id);
+                            result = "ok";
+                        }
+                    } catch (NumberFormatException e) {
+                        result = "Введите коректные данные";
+                    }
+                }
+                String json = new Gson().toJson(result);
+                PrintWriter pw = new PrintWriter(new OutputStreamWriter(resp.getOutputStream(), "UTF-8"));
+                pw.append(json);
+                pw.flush();
+            }
         }
     }
